Changed to unsigned 15 bit, create the memory read/write logic, removed particular cases (perfect input file), removed copy paste code and created method, accesul la memory location prin registru

+++ ProjectSQ/Models/Memory.cs
@@ -6,7 +6,7 @@ namespace ProjectSQ.Models
     {
         public static int currentInstruction = 0;
         public static string[] internalMemory = new string[1024];
-        public static short[] programData = new short[65536];//0->60.000 mem, restul stack
+        public static byte[] programData = new byte[65536];//0->60.000 mem, restul stack
         public static int instructionNumber = 0;
 
         //program's configuration file
